before jest done

+++ Part4/bloglist/controllers/blogs.js
@@ -3,7 +3,7 @@ const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
 
 blogsRouter.get('/',async(request, response) => {
-  const blogs = await Blog.find({})
+  const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 })
   response.status(200).json(blogs)
 })
 
@@ -18,10 +18,17 @@ blogsRouter.get('/:id', async (request, response) => {
 
 blogsRouter.post('/',async (request, response) => {
   const body = request.body
+
+  const user = request.user
+
+  if (!user) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
   const blog = new Blog({
     title: body.title,
     author: body.author,
     url: body.url,
+    user:user.id,
     likes: body.likes
   })
   if (!blog.title || !blog.url) {
@@ -32,13 +39,27 @@ blogsRouter.post('/',async (request, response) => {
       blog.likes = 0
     }
     const savedBlog = await blog.save()
+    user.blogs = user.blogs.concat(savedBlog._id)
+    await user.save()
     response.status(201).json(savedBlog)
   }
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndDelete(request.params.id)
-  response.status(204).end()
+  const user = request.user
+
+  if (!user) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
+
+  const blog = await Blog.findById(request.params.id)
+  if (blog.user.toString() === request.user.id) {
+    await Blog.findByIdAndDelete(request.params.id)
+    response.status(204).end()
+  } else {
+    return response.status(401).json({ error: 'Unauthorized to delete the blog' })
+  }
+
 })
 
 blogsRouter.put('/:id',async (request, response) => {
